[#10] refactor: 섹션 6. Service (서비스)
- Service로 모든 로직 옮기기
